grid workspace layout: implement actions to switch in grid
Actions GoToDesktop & SendToDesktop named "left" and "right" are now performed just
on the same row. In case that wrap is True, the switch happens to the
workspace on the other side of the row.

Similarly it's newly possible to perform similar switches "up" and
"down", where actions are performed just in area of the same column.

The grid is created virtually on top of wl_list for workspaces provided
by wlroots. So this solution is compatible with current protocols.
The logic is simple:
  * calculate index of target workspace
  * go through the list of workspaces until the required index is
    discovered
It's less effective in comparison to previous implementation of
"left" and "right" operations, however I do not expect anyone to
create more than 25 workspaces and that's quite short list to
go through so let's survive it.